web/incomplete/team_dashboard/requirements.txt
fastapi==0.120.1
uvicorn==0.38.0
sqlalchemy==2.0.44
pydantic==2.12.3
passlib==1.7.4
psycopg2==2.9.11
argon2
argon2-cffi==25.1.0
pydantic-settings
pydantic[email]
python-multipart
python-jose[cryptography]
jinja2
